requirements.txt
fastapi
uvicorn
faiss-cpu
sentence-transformers
transformers
torch
numpy
pydantic
pandas
scipy
accelerate>=0.26.0
openai
supabase
python-dotenv